+++ packages/tdesign-uniapp/global.d.ts
@@ -45,6 +45,7 @@ declare module 'vue' {
     TNavbar: typeof import('@tdesign/uniapp/navbar/navbar.vue').default;
     TNoticeBar: typeof import('@tdesign/uniapp/notice-bar/notice-bar.vue').default;
     TOverlay: typeof import('@tdesign/uniapp/overlay/overlay.vue').default;
+    TParagraph: typeof import('@tdesign/uniapp/paragraph/paragraph.vue').default;
     TPicker: typeof import('@tdesign/uniapp/picker/picker.vue').default;
     TPickerItem: typeof import('@tdesign/uniapp/picker-item/picker-item.vue').default;
     TPopover: typeof import('@tdesign/uniapp/popover/popover.vue').default;
@@ -77,7 +78,9 @@ declare module 'vue' {
     TTable: typeof import('@tdesign/uniapp/table/table.vue').default;
     TTabs: typeof import('@tdesign/uniapp/tabs/tabs.vue').default;
     TTag: typeof import('@tdesign/uniapp/tag/tag.vue').default;
+    TText: typeof import('@tdesign/uniapp/text/text.vue').default;
     TTextarea: typeof import('@tdesign/uniapp/textarea/textarea.vue').default;
+    TTitle: typeof import('@tdesign/uniapp/title/title.vue').default;
     TToast: typeof import('@tdesign/uniapp/toast/toast.vue').default;
     TTreeSelect: typeof import('@tdesign/uniapp/tree-select/tree-select.vue').default;
     TUpload: typeof import('@tdesign/uniapp/upload/upload.vue').default;